requirements.txt
langchain==0.0.154
python-dotenv==1.0.0
streamlit==1.28.0
faiss-cpu==1.7.4
streamlit-extras
altair<5
pytesseract
pdf2image
openai==0.28.1
tiktoken==0.5.1
